use parseFloat

+++ src/views/reports/DashboardView/EarningsSegmentation/index.js
@@ -195,7 +195,7 @@ const EarningsSegmentation = ({ className, ...rest }) => {
       }
     },
     // series: [(100 * parseInt(stateDevice.pcs[0].sInP1)/(parseInt(stateDevice.pcs[0].acInAcP)+parseInt(stateDevice.pcs[0].sInP1))).toFixed(1)]
-    series: [(100 * parseInt(stateDevice.pcs[0].greenPer)).toFixed(1)]
+    series: [(100 * parseFloat(stateDevice.pcs[0].greenPer)).toFixed(1)]
   };
 
   const DischargeButton = withStyles((theme) => ({
